fix forwards compat tests

+++ chia/wallet/cat_wallet/cat_wallet.py
@@ -209,15 +209,20 @@ class CATWallet:
         self.lineage_store = await CATLineageStore.create(self.wallet_state_manager.db_wrapper, self.get_asset_id())
         await self.wallet_state_manager.add_new_wallet(self, self.id())
 
+        delete: bool = False
         for state in await self.wallet_state_manager.interested_store.get_unacknowledged_states_for_asset_id(
             limitations_program_hash
         ):
-            peer_id: bytes32 = (await self.wallet_state_manager.wallet_node.get_full_node_peer()).peer_node_id
-            await self.wallet_state_manager.retry_store.add_state(state[0], peer_id, state[1])
+            new_peer = await self.wallet_state_manager.wallet_node.get_full_node_peer()
+            if new_peer is not None:
+                delete = True
+                peer_id: bytes32 = new_peer.peer_node_id
+                await self.wallet_state_manager.retry_store.add_state(state[0], peer_id, state[1])
 
-        await self.wallet_state_manager.interested_store.delete_unacknowledged_states_for_asset_id(
-            limitations_program_hash
-        )
+        if delete:
+            await self.wallet_state_manager.interested_store.delete_unacknowledged_states_for_asset_id(
+                limitations_program_hash
+            )
 
         return self
